ns: rewrite IDownloadTaskInterface
